+++ run.py
@@ -1,4 +1,4 @@
-from application import Application
+from app.application import Application
 
 if __name__ == "__main__":
     app = Application()
